requirements.txt
appnope==0.1.3
asttokens==2.4.1
black==23.12.1
click==8.1.7
comm==0.2.1
contourpy==1.2.0
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
executing==2.0.1
fonttools==4.47.0
ipykernel==6.28.0
ipython==8.19.0
jedi==0.19.1
jupyter_client==8.6.0
jupyter_core==5.7.0
kiwisolver==1.4.5
matplotlib-inline==0.1.6
mypy-extensions==1.0.0
nest-asyncio==1.5.8
numpy==1.26.3
opencv-contrib-python==4.7.0.68
opencv-python==4.7.0.68
packaging==23.2
parso==0.8.3
pathspec==0.12.1
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.1.0
prompt-toolkit==3.0.43
psutil==5.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
pyntcore==2024.1.1.1
pyparsing==3.1.1
python-dateutil==2.8.2
pyzmq==25.1.2
robotpy-wpimath==2024.1.1.1
robotpy-wpinet==2024.1.1.1
robotpy-wpiutil==2024.1.1.1
six==1.16.0
stack-data==0.6.3
tornado==6.4
traitlets==5.14.1
wcwidth==0.2.12
